+++ package.json
@@ -23,16 +23,16 @@
   },
   "homepage": "https://github.com/qavajs/playwright-wdio#readme",
   "devDependencies": {
-    "@types/express": "^5.0.3",
-    "@types/node": "^24.6.2",
+    "@types/express": "^5.0.5",
+    "@types/node": "^24.10.1",
     "express": "^5.1.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3"
   },
   "dependencies": {
-    "@playwright/test": "^1.55.1",
-    "@qavajs/memory": "^1.10.2",
-    "@qavajs/playwright-runner-adapter": "^1.6.0",
+    "@playwright/test": "^1.56.1",
+    "@qavajs/memory": "^1.10.3",
+    "@qavajs/playwright-runner-adapter": "^1.8.0",
     "@qavajs/playwright-wdio-fixtures": "^0.0.4",
     "webdriverio": "^9.20.0"
   }
