+++ src/poDefine.ts
@@ -1,5 +1,5 @@
 import { When } from '@qavajs/playwright-runner-adapter';
-import { type MemoryValue } from './types';
+import type { MemoryValue } from './types';
 import { locator } from './pageObject';
 
 /**
