+++ steps.js
@@ -9,6 +9,6 @@ require('./lib/mock.js');
 require('./lib/poDefine.js');
 require('./lib/mouse.js');
 require('./lib/keyboard.js');
-require('./lib/utils.js');
+require('./lib/waits.js');
 require('./lib/http.js');
 require('./lib/files.js');
